refactor(gda-balancing): isolate operation conformance harness
Separate production and independent Operation execution adapters, then centralize vector discovery, canonical observations, divergence reporting, and candidate-graph gates in a test-only orchestrator.
